requirements.txt
matplotlib==3.10.1
numpy==2.2.4
streamlit==1.43.2
Pillow==8.4.0